refactor : 한번에 여러 번 이동할 수 있는 기물들에 대하여 추상화하여 변경
